+++ lib/src/commands/do.dart
@@ -34,7 +34,6 @@ class Do extends Command<void> {
     addSubcommand(deps.doCommit);
     addSubcommand(deps.doPush);
     addSubcommand(deps.doPublish);
-    addSubcommand(deps.doConfigurePublish);
     addSubcommand(deps.doUpgrade);
     addSubcommand(deps.doCheckout);
   }
@@ -50,15 +49,12 @@ class DepsOfDo {
     DoCommit? doCommit,
     DoPush? doPush,
     DoPublish? doPublish,
-    DoConfigurePublish? doConfigurePublish,
     DoUpgrade? doUpgrade,
     DoCheckout? doCheckout,
   }) : create = create ?? Create(ggLog: ggLog),
        doCommit = doCommit ?? DoCommit(ggLog: ggLog),
        doPush = doPush ?? DoPush(ggLog: ggLog),
        doPublish = doPublish ?? DoPublish(ggLog: ggLog),
-       doConfigurePublish =
-           doConfigurePublish ?? DoConfigurePublish(ggLog: ggLog),
        doUpgrade = doUpgrade ?? DoUpgrade(ggLog: ggLog),
        doCheckout = doCheckout ?? DoCheckout(ggLog: ggLog);
 
@@ -76,9 +72,6 @@ class DepsOfDo {
 
   /// The do publish command
   final DoPublish doPublish;
-
-  /// The do configure-publish command
-  final DoConfigurePublish doConfigurePublish;
 
   /// The do upgrade command
   final DoUpgrade doUpgrade;