+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - refactor(gg_one): remove can merge from the CLI
 - refactor(gg_one): remove can upgrade from the CLI
 - refactor(gg_one): remove the check command group from the CLI
+- refactor(gg_one): remove do configure-publish from the CLI
 
 ## 12.5.0 - 2026-08-02